Insert Data to Database - Shorter Methods [ FreeCodeCamp did not accept due to preset rules ]. But Works
This script is not accepted during execution with FreeCodeCamp Server due to preset rules on the scripting process, But this script has the same shorter methods incorporated as the insert_data.sh file used for the FreeCodeCamp Server, With some more added methods incorporated with the variables in the first few lines which can run on any server "which does not have preset rules in the execution process of the script". [ Compare Both Files to Identify the Difference. ]