Clean and Modify Data in the Database using Script - A Self Endeavour
A Self Endeavor And Scope Beyond the Project Goals, This Script Automates the Process of  Creating and Dropping of Tables,  Rows  and Columns, Cleaning and Modifying Data along with Setting Constraints. Rather than Manually Hard Coding using psql Command Prompt which was requested by the FreeCodeCamp Curriculum.

Key factor's to Note :

To Minimize code array with loops are incorporated, But to achieve minimalism multi-dimensional arrays are required but due to the fact 
bash is subject to single dimension array, A work around is incorporated by using POSIX shell built-in command " eval "

NOTE:  To make the code more minimal one could incorporate JSON  instead of Array.